rag pipeline service implemented for tests and web app functionality
