+++ src/AI/api/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+api_blueprint = Blueprint('ai', __name__)
+
+from . import handle_prompt
