+++ src/AI/langchain_methods/__init__.py
@@ -1,0 +1,1 @@
+from .rag_pipeline import rag_pipeline
