+++ src/AI/langchain_methods/rag_pipeline.py
@@ -1,28 +1,9 @@
-import os
-from dotenv import load_dotenv
+from utils import decode_base64_image
+from fashion_search import TextToImageSearch, ImageToImageSearch
 from langchain.chains import LLMChain
 from langchain.prompts import PromptTemplate
 from langchain_google_genai import GoogleGenerativeAI
-from langchain.memory import ConversationBufferMemory
-from fashion_search.text_to_image import TextToImageSearch  
-from fashion_search.image_to_text import ImageToTextGenerator
-from fashion_search.image_to_image import ImageToImageSearch
-import io
-from PIL import Image
-import base64
-from time import sleep
-
-
-load_dotenv()
-
-
-def decode_base64_image(image_base64: str) -> Image.Image:
-    """Decode a base 64 encoded image string and return Pil iamge object"""
-    image_base64 = image_base64.split(",")[1]
-    image_data = base64.b64decode(image_base64)
-    image = Image.open(io.BytesIO(image_data)).convert('RGB')
-    return image
-
+import os
 
 def rag_pipeline(query_text, category, image_base64=None, memory=None):
     """Retrieval-Augmented Generation (RAG) pipeline using LangChain with conversation history."""
@@ -88,33 +69,3 @@ def rag_pipeline(query_text, category, image_base64=None, memory=None):
     response = chain.run(context=context_str, query_text=query_text, image_base64= image_base64)
     
     return response
-
-
-if __name__ == "__main__":
-    
-    GOOGLE_API_KEY = os.getenv("google_api_key")
-    if not GOOGLE_API_KEY:
-        raise ValueError("Google API key is missing! Ensure it's set in the .env file.")
-    
-
-    memory = ConversationBufferMemory(memory_key="chat_history", input_key="query_text", return_messages=True)
-    
-
-    query = "I need a yellow dress for a summer night out. It should have floral prints. Please consider the uploaded image as well."
-    category = "clip_DRESSES_JUMPSUITS"  
-    
-    f = open("ornek.txt", "r")
-    image_base64 = f.readline()
-    print(image_base64)
-    
-    
-    recommendation = rag_pipeline(query, category, image_base64, memory)
-    print("Recommendation:\n", recommendation)
-    
-    
-    sleep(60)
-    
-    #   follow-up query with chat history included
-    followup_query = "I liked that suggestion, but can you recommend a piece with more flowers?"
-    followup_recommendation = rag_pipeline(followup_query, category, image_base64, memory)
-    print("Follow-up Recommendation:\n", followup_recommendation)
