+++ src/AI/utils/__init__.py
@@ -1,0 +1,1 @@
+from .decode_base64_image import decode_base64_image
